restructure patch directory, begin to use type-specific files
